# Optimize for speed