Here are some boot time and memory usage for the **2nd boot** (1st one being ignored due to cloud-init does its thing) of a guest VM.

From:

```
# systemd-analyze 
Startup finished in 780ms (kernel) + 7.248s (userspace) = 8.029s 
graphical.target reached after 5.743s in userspace
```

```
root@u22:~# free -mt
               total        used        free      shared  buff/cache   available
Mem:            3932          94        3614          16         222        3777
Swap:              0           0           0
Total:          3932          94        3614
```

To:

```
# systemd-analyze 
Startup finished in 522ms (kernel) + 2.729s (userspace) = 3.252s 
graphical.target reached after 2.709s in userspace
```

```
# free -mt
               total        used        free      shared  buff/cache   available
Mem:            3932          82        3674          16         175        3792
Swap:              0           0           0
Total:          3932          82        3674
```

This helps speeding up the boot of the nested `micro0X` VMs even if the total test runtime seems only slightly reduced.

The biggest test runtime gain is from `test/includes/microcloud: don't start VMs after snapshot, reset_systems will`'s commit that speeds up the `setup` phase (~9.5 minutes down to ~7.5 minutes).