@simondeziel 2024-05-02
> @masnax I think you've been working in this area, any idea why those tests now all fail early with:
> 
> ```
> Error: System "micro02" failed to join the cluster: Failed to update cluster status of services: Failed to join "LXD" cluster: Failed to configure cluster :Failed request to add member: The joining server version doesn't match (expected 5.21.1 with API count 385)
> ```

Was a one time hiccup it seems. Sorry for the ping.