@simondeziel 2024-05-02
Yay, it seems the longest test consistently finishes in around 3.5h. Not great when taken at face value but it used to hit the timeout limit of 6 hours pretty often before, for me at least.